不要なファイルの削除、実行ファイルの更新
