+++ 06.bowling_object/frame.rb
@@ -3,15 +3,55 @@
 require './shot'
 
 class Frame
-  attr_reader :first_shot, :second_shot, :third_shot
-
-  def initialize(first_mark, second_mark = nil, third_mark = nil)
-    @first_shot = Shot.new(first_mark)
-    @second_shot = Shot.new(second_mark)
-    @third_shot = Shot.new(third_mark)
+  FRAME_RANGE = 0..9
+  def initialize(shots)
+    @shots = shots
   end
 
-  def score
-    @first_shot.score + @second_shot.score + @third_shot.score
+  def generate_point_cal_frame
+    frames = generate_frames
+    point_cal_frame = []
+
+    FRAME_RANGE.each { |n| point_cal_frame << generate_cal_frame(frames, n) }
+
+    point_cal_frame
+  end
+
+  private
+
+  def generate_frames
+    frame = []
+    frames = []
+    @shots.each do |shot|
+      frame << shot
+      frames = generate_frame(shot, frame, frames)
+      frame.clear if frame.size >= 2 || shot == 10
+    end
+    frames
+  end
+
+  def generate_frame(shot, frame, frames)
+    if frames.size < 10
+      frames << frame.dup if frame.size >= 2 || shot == 10
+    else # last frame
+      frames.last << shot
+    end
+    frames
+  end
+
+  def generate_cal_frame(game, num)
+    frame, next_frame, after_next_frame = game.slice(num, 3)
+    next_frame ||= []
+    after_next_frame ||= []
+    left_shots = next_frame + after_next_frame
+
+    if num == 9 # last
+      return [frame[0], frame[1], frame[2]] if frame[0] == 10 || frame[0] + frame[1] == 10 # strike or spare
+    else
+      return [frame[0], left_shots[0], left_shots[1]] if frame[0] == 10 # strike
+      return [frame[0], frame[1], left_shots[0]] if frame.map.sum == 10 # spare
+    end
+
+    [frame[0], frame[1]]
   end
 end
